feat: create DefaultNavLogo
