+++ frontend/src/components/layout/LogoComponents/index.ts
@@ -1,0 +1,3 @@
+export { DefaultHeaderLogo } from './DefaultHeaderLogo';
+export { ConfigurableLogo } from './ConfigurableLogo';
+export { DefaultNavLogo } from './DefaultNavLogo';
